In train.py: Added Preprocessing function, logic for RL loop and (commented out) visualization code. The folders at the end mentioned are written to for Saver() models, monitor files that store gameplay videos & visualization through tensorboard.
In agent.py: Implemented initial version of agent which uses categorical dist. with 51 bins. As needed, Implements the DQN tricks like target networks and an experience replay as well. The entire algorithm logic for the optimization flows through a Tensorflow graph with the final computed target being passed through stop_gradient, cross-entropy between the training net's action-value distribution of the action taken and the target dist (with the distributional bellman operator).